refactor: move withTimeout implementation to utils
