+++ lib/utils.ts
@@ -199,16 +199,22 @@ export function isValidExchangeRate(rate: number | null | undefined): rate is nu
   return rate != null && rate > 0 && rate < 100000
 }
 
-/**
- * Race `promise` against a deadline.  Rejects with a `TimeoutError` if the
- * promise has not settled within `ms` milliseconds.  Useful for bounding
- * network calls that would otherwise hang the render path.
- */
-export function withTimeout<T>(promise: Promise<T>, ms: number): Promise<T> {
-  return Promise.race([
-    promise,
-    new Promise<T>((_, reject) =>
-      setTimeout(() => reject(new Error(`Timed out after ${ms}ms`)), ms),
-    ),
-  ])
+// Run a promise against a wall-clock budget. The underlying work continues to
+// completion on the server when the budget elapses: we just stop waiting for
+// it. For Anthropic calls that's fine: a slow Opus turn finishing later still
+// warms its own cache.
+function withTimeout<T>(promise: Promise<T>, ms: number): Promise<T> {
+  return new Promise<T>((resolve, reject) => {
+    const timer = setTimeout(() => reject(new Error(`Timeout after ${ms}ms`)), ms)
+    promise.then(
+      (v) => {
+        clearTimeout(timer)
+        resolve(v)
+      },
+      (e) => {
+        clearTimeout(timer)
+        reject(e)
+      },
+    )
+  })
 }
